requirements.txt
playwright>=1.40.0
openpyxl>=3.1.0
requests>=2.28.0
pyyaml>=6.0
flet>=0.21.0
